requirements.txt
streamlit
streamlit_folium
folium
watchdog
plotly
python-dotenv
psycopg2-binary==2.9.11
black

numpy==1.26.4
scikit-learn==1.4.2
xgboost==1.7.3

numba==0.59.1
llvmlite==0.42.0
shap==0.45.0
numpy==1.26.4
pyarrow==15.0.2

matplotlib
ipykernel
